Changing the @OutputFile on LessTask to a @OutputDirectory
Changing the defined output of the LessTask from Directory to
File. The variable is treated as a directory (i.e. sent in as
a first param in the two param File constructor) and using the
OutputFile annotation causes gradle to do some funky stuff and
error out with something like 'cannot write to output X because
it is a directory'.

This also enables gradle output directory auto creation.