Changing less task name to lesscss due to name conflict
The LessExtension is registered with the project with the name
'less'. This means that writing:

less {
  source = x
  dest   = y
}

is actually _not_ configuring the task but is instead just
setting some variables on the 'less' extension. This can be
proven by doing a 'println less.class.name' which prints out
x.y.z.LessExtension_Decorated and not x.y.z.LessTask_Decorated.

I have left the extension with the name 'less' and changed the
task name (and references in the plugin and test code) to
'lesscss'. This seems to work better, the plugin.gradle code
seems to do what it is supposed and my project using the task
can now successfully call the task with expected results.